validacion rol

+++ ListarAcudientesDB.php
@@ -23,8 +23,28 @@ if(!($sesion==null) && !($sesion==null)){
 </head>
 <body>
 	<div>
+	<?php
+	if($rol==1){
+	?>
 	<button><a href="inicioAdmin.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Volver</a></button>
+	<?php
+	}
+	?>
+		<?php
+	else{
+	?>
+	<button><a href="inicio.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Volver</a></button>
+	<?php
+	}
+	?>
+	
+	<?php
+	if($rol==1){
+	?>
 	<button><a href="add_acudienteDB.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Agregar acudiente</a></button>
+	<?php
+	}
+	?>
 	<table>
 		<tr>
 			<th>Id acudiente</th>
@@ -34,7 +54,13 @@ if(!($sesion==null) && !($sesion==null)){
 			<th>Dirección</th>
 			<th>Teléfono</th>
 			<th>Id alumno</th>
+			<?php
+			if($rol==1){
+			?>
 			<th>Opciones</th>
+			<?php
+			}
+			?>
 		</tr>
 		<?php
 		foreach($personas as $p){
@@ -47,7 +73,13 @@ if(!($sesion==null) && !($sesion==null)){
 				<td><?=$p->direccion;?></td>
 				<td><?=$p->telefono;?></td>
 				<td><?=$p->estudiante_id_alumno;?></td>
+				<?php
+				if($rol==1){
+				?>
 				<td><button><a href="edit_acudienteDB.php?id=<?=$p->id;?>&estudiante_id_alumno=<?=$p->estudiante_id_alumno;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Editar acudiente</a></button><br><button><a href="eliminar_acudienteDB.php?id=<?=$p->id;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Eliminar</a></button></td>
+				<?php
+				}
+				?>
 			</tr>
 			<?php
 		}
