+++ ListarMatriculasDB.php
@@ -23,8 +23,29 @@ if(!($sesion==null) && !($sesion==null)){
 </head>
 <body>
 	<div>
+			<?php
+	if($rol==1){
+	?>
 	<button><a href="inicioAdmin.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Volver</a></button>
+	<?php
+	}
+	?>
+		<?php
+	else{
+	?>
+	<button><a href="inicioAdmin.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Volver</a></button>
+	<?php
+	}
+	?>
+	
+		<?php
+	if($rol==1){
+	?>
 	<button><a href="add_matriculaDB.php?sesion=<?=$sesion?>&rol=<?=$rol?>">Agregar matrícula</a></button>
+	<?php
+	}
+	?>
+	
 	<table>
 		<tr>
 			<th>Id matrícula</th>
@@ -32,7 +53,14 @@ if(!($sesion==null) && !($sesion==null)){
 			<th>Año lectivo</th>
 			<th>Calendario</th>
 			<th>Estado</th>
-			<th>Opciones</th>
+				<?php
+	if($rol==1){
+	?>
+	<th>Opciones</th>
+	<?php
+	}
+	?>
+			
 		</tr>
 		<?php
 		foreach($personas as $p){
@@ -43,7 +71,14 @@ if(!($sesion==null) && !($sesion==null)){
 				<td><?=$p->ano_lectivo;?></td>
 				<td><?=$p->calendario;?></td>
 				<td><?=$p->estado;?></td>
-				<td><button><a href="edit_matriculaDB.php?idMatricula=<?=$p->idMatricula;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Editar matrícula</a></button><br><button><a href="eliminar_matriculaDB.php?idMatricula=<?=$p->idMatricula;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Eliminar</a></button></td>
+					<?php
+	if($rol==1){
+	?>
+	<td><button><a href="edit_matriculaDB.php?idMatricula=<?=$p->idMatricula;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Editar matrícula</a></button><br><button><a href="eliminar_matriculaDB.php?idMatricula=<?=$p->idMatricula;?>&sesion=<?=$sesion?>&rol=<?=$rol?>">Eliminar</a></button></td>
+	<?php
+	}
+	?>
+				
 			</tr>
 			<?php
 		}
